move gyro headers to gyro folder
